fix: navigate to main screen many time if already authen

+++ lib/screens/dashboard_screen/widgets/products_pie_chart.dart
@@ -101,7 +101,7 @@ class ProductPieChartState extends State<ProductsPieChart> {
       final value = data[i]["value"];
       return PieChartSectionData(
           color: data[i]["color"],
-          value: value,
+          value: value.toDouble(),
           title: '${value / widget.totalSoldCount * 100}%',
           radius: radius,
           titleStyle: AppStyles.bodySmall.copyWith(
